Significant refactor half way toward initiating TLS
There were some challenges which warranted a bit of refactoring here.
Most importantly, tokio_rustls::TlsConnector requires a DNS name to
connect to, probably for certificate validation and SNI and stuff. That
means that we have to have that name somewhere ...

We could've put that name into the ConnectionWorker, however, that means
that there are a lot more failure cases which cannot be predicted by the
starttls() method, leading to errors which are swallowed.

To avoid these, we pulled the entire TLS configuration state into the
ConnectionHandle instead. The downside is that the handle becomes
larger, but I suppose the worker would've become larger instead, so...
At least the string and context and stuff get dropped after starttls.