Do not GC on each event loop iteration
Now we only GC after a registry key has truly been dropped. This is
signalled using a flag and an optional write to the event queue: the
write to the event queue is omitted if at least one event is in there.
That is sound because the atomic write to the flag happens before the
check for capacity, which means (together with SeqCst) that the
mainloop will have to iterate at least once after the flag has been set.

This should improve performance somewhat and even significantly in
settings where a lot of handles are open.