Introduce block_reads() API method
This is to be used in conjunction with StartTLS initiation. It prevents
a common race condition in the following steps:

1. Client requests to start TLS
2. Server sends acknowledgement
3. Server enters TLS mode
4. Client starts TLS

If 4 happens before 3 (which is possible as 4 only depends on 2), the
TLS handshake will be treated as plain text data and handed to Lua. We
don't want that.

Unfortunately, this means that changes to all Lua starttls() call sites
are required.