Expose the bare minimum of the cert for util.x509 to work
In the future, we'd want to actually refactor things so that the network
backend can do all the validation. The Lua-based backends would continue
to use util.x509, the rust based backend would do its own thing. The
reference identifiers would be passed as `extra` to the sockets in
net.connect or during starttls().