Add support for arbitrary keys on connections
This sucks, but it's currently required for net.http, which stores
request metadata on the connection handle.

Better ways of doing this would include somehow reaching below mlua and
gaining access to the table it is *already* using (in some lua versions)
as user value, to avoid at least one level of nesting. Another good
thing would be to expose the possibility of having multiple user values
in Lua 5.4+.