Get rid of LuaError::RuntimeError use
It is not for us. We are supposed to use ExternalError.

However, in some cases we really don't have a proper Rust error to
expose. For those, we use a new OpaqueError type which is really just a
string. It isn't useful for error handling in Rust, but we can still use
it to raise an error to Lua.