Introduce full duplex operation on AsyncRead + AsyncWrite
This allows us to do actual full duplex on streams, without having to
rely on a lock (as tokio::io::split does). This works because we know
that we are running this from the same task, always.

It also allows us to avoid returning from the future while writes are
going on in the background and only return when either the write buffer
is exhausted or a read has completed.