Fix debug assert which will always trigger
remaining() is certain to be zero at that place, because read_buf told
us it had not received any bytes.

What we *truly* want to look at is remaining_mut(), to detect when we
accidentally pass an empty buffer to read_buf even though we expect to
read more data.