Refactor stream wrapping
This has been a major pain to look at for a while now. It was one of the
oldest parts of the code and it had started to smell. This is hopefully
a bit cleaner in separation, focusing on the Lua stuff at the right
places and keeping it out of others.

In addition, this finally exposes the correct port numbers to Prosody.