Improve memory and allocation footprint
As it turns out, rust coroutines have to keep all possible variants of
their stack on the heap. So far so obvious. However, this grows in
considerable ways when you consider nested coroutine calls.

In order to reduce this overhead, we can avoid unnecessary coroutine
wrappings, such as in the tokio::spawn cases.

A more trivial thing is to use a reference instead of box in the inner
loop to avoid a needless shortlived allocation.